use two different values for "outside" and "border"
This commit changes the type of the heightmap from uint8_t to int16_t, using a
typedef so that changing it again is no problem.

That change was motivated by the need to use a different value for outside and
border cells. Originally, -1 and 0 were chosen, but that breaks the divergence
calculation (which was wrongly assuming that outside cells were valid).

In order to keep that calculation simple, now:
 * outside cells have a value of 0
 * border cells have a value of 1

Note that a (very good) side-effect of this commit is that now border cells are
rightly considered in the divergence calculation. Only outside cells are
ignored.