more test of skeleton extraction
The simple_corridor_set test has been extended so that the
corridor shape has a "growing then shrinking region". This
shows a normal behaviour of the heightmap algorithm, in which
edges from the skeleton to the border are found.

This would have to be filtered at the graph construction level.

Moreover, the multipath test is revived.

Note that currently both tests are very verbose and do not
actually check anything automatically.