better Voronoi diagram extraction algorithm
After having read the litterature, Jochen's method
of "height maps" is actually a common one for
raster-based Voronoi diagram extraction.

As I expected, it is perfectly possible to extract
the diagram in one simple pass. This is done now. In
a normal Voronoi diagram creation, the following is done
for each "object":

 * each step, we propagate the value to higher nodes
 * if neighbour nodes are either the same value or the same
   propagated value, we have a point of the diagram.

In our case, this is invalid as the "objects" are the boundary
points (and therefore we would get the voronoi diagram between all
boundary points, which is more or less the corridor interior).
What we want is the following:
 * the voronoi diagram of the boundary sets (i.e. the objects
   made of connected points)
 * an association between each point of the diagram and each
   point of the boundary.

To do both, we still build the diagram as if each boundary
point is its own object, but we ignore the diagram points where
the associated boundary points are adjacent. This is made by tracking
which boundary point generates which distanaces, and therefore
we also get the diagram point <=> boundary point association.