fix fractional distance issues in the creation of the Voronoi diagram
In 2e3320695af9563cd1cc120ee332696fe0a6da78, we switched to a
method where a point is added to the Voronoi set if
 * a hill is detected
 * the two points generators are not connected (i.e. they
   are not generated by closeby points of the border)

This commit fixes issues with the algorithm that was throwing away
possible Voronoi candidates:

First, to properly keep track of the association with border point, we
have to propagate parents more often than what was done. When

   cost(current) + edge_cost == cost(neighbour)

then parents(current) should also be registered as parents of
neighbour.

Because of that, another problem has arisen in the determination
of "neighbouring parents". What we were doing is ignoring a
candidate to the Voronoi set when any parent of current touch
any parent of neighbour. This is wrong because of the following:
 - let's imagine that both current and neighbour are Voronoi points
 - current is already registered as Voronoi point and therefore is
   associated with at least two disjoint parts of the border
 - neighbour is still associated with only one part of the border

With the old method, +neighbour+ would be thrown away as being adjacent
to +current+. And with the parent propagation change mentionned earlier,
+neighbour+ will never be added to the Voronoi set.

We fix that by maintaining the sets of connected border points: each
possible candidate is described by a MedianPoint struct which maintains
the border points in separate sets (see documentation of
MedianPoint::borders). That way, the "throwing away" condition is changed
into:

  a point should not be added to the Voronoi diagram if any
  of its borders have a connection with a border of the other
  point.