much better implementation of the graph building algorithm
In this new implementation, we progress through the median
points by going from neighbour to neighbour, building the
corridor set incrementally. The criteria to add a point in
the corridor is the same (i.e. that the borders to the point
are adjacent to the borders of the corridor).

What changes is that, now, a point is added to the corridor
set only if both its borders and itself are adjacent to the
corridor. The previous algorithm would, in the end, merge
two corridors if a third one was joining them.