fix: merging as we do it an assymetric operation
The general merging algorithm was as follows:

  for each element "p" in left
     if p can be merged
        merge
     else
        push(p)
        push(element of right that is closest to p)
     end
  end

There is obviously a problem as all elements of right are not
tested *and* it is possible (for instance in perpendicular corridors)
that a lot of elements of left have the same closest element in
right.

This commit fixes the algorithm by doing it two-pass:

  for each element "p" in left
     if p can be merged
        merge
     else
        push(p)
     end
  end

  for each element "p" in right
     if p has been merged in first pass
        do nothing
     else
        push(p)
     end
  end

Et voila ...

Last thing to do: actually connect the parts generated by the merge.