finish the corridor merging algorithm by adding connections
we only add connections between non-merge corridors and merge
corridors. Otherwise, we would be creating new path that were
in none of the plans, which is not the intent.

The intent is really to represent how one plan is related to
the other, not to merge them as "using both solutions at the
same time".