skel: properly consider points with more than 2 borders
Currently, we were using all points to be the same. This introduced
an assymetry in the plan building process, since median points with
more than 2 borders should actually be shared by multiple corridors.

Given that sharing a point is not supported in our model (the merge
algorithm would not support it properly), we instead remove them from
the resulting set of median points. Nonetheless, we still use them to
properly update the connections between corridors.