provide a generic implementation for building connected point sets
the new method, Nav::updateConnectedSets, can be used to incrementally
build a list of point sets so that no two sets in the list are touching
each other, and all elements of one set form a connected set.

This is a very dumb implementation, but quite enough given the size
of the point sets we have to manage.