bin: improve bin/nav
This commit improves bin/nav and the general usefulness of the various
classes which form the frontend of the navigation package
(i.e. DStar and TraversabilityMap).

In practice, the interfaces are changed so that it is possible to
specify a mapping from a 8-bit value into both a 4-bit traversability
class and a floating-point speed (in m/s). The cost for DStar is then
computed as the time to travel. Moreover, the associated methods to load
everything from files (GDAL files for the map and a text file for the
classes) is also provided.

bin/nav is updated to use these capabilities, and now generates images
(using GDAL) instead of cryptic text files.