in SE::buildPlan, fix handling of crossroad points
Until now, we were testing the connectivity of corridors to
crossroad points using isBorderAdjacent, which is the
theoretical method.

In practice, in complex zones of the image, it is possible that
a (theoretical) high-connectivity crossroad point is in practice
multiple points of lower connectivity.

In practice, it is enough to build the connected sets of the
crossroad points (i.e. what crossroad points touch each other),
and test if the corridor median line touches one of these sets.