cleanup MedianPoint and MedianLine
Until now, the MedianPoint class was not including the associated
center point, which was stored instead in MedianLine. This is
obviously stupid, and a +center+ field is added to MedianPoint
for this information.

Moreover, MedianLine is now a list<> as the map<> functionality
was seldom used *and* was a big performance hit.

Moreover, the long dead buildCrossroads() and findAdjacentBorders
are not completely removed from here.