improve D* performance by using a pool allocator
One problem with the initial run of D* is that a lot of
insert/erase operations are done on the open list, which
in our implementation does a lot of new/delete (because
map<> and multimap<> are used).

This commit adds a memory pool allocator and uses it in the
D* implementation, yielding a 15% performance improvement.