problem in adding additional right points as merged
Right now, the point mapping from right to left (i.e. how
we deal with the assymetry of the algorithm) is problematic
because it associates some points with the wrong merged point.

Trying to change the method: the new associations should be
done in the second pass (when we deal with the right
corridors) instead.