reimplemented the corridor extraction using a DFS-based approach
This approach gives very good results (faster, simpler).

Also: removes the use of tuples for connection representation,
replaced by a structure. Moreover, since the DFS gives us at
which side the corridors are connected, we also removed PointID
in the connection descriptor, replaced by a "side" parameter
(whether the connection comes from/goes to the front or the back)