don't create a link between two corridors if the two connecting points are from the same corridor
This commit fixes issues with two-pixel corridors.

Let's assume such a two pixel corridor was part of the crossroad
c0 (on its front side) and c1 (on its back side). Because it is a
two-pixel corridor, the two endpoints are touching each other
and therefore c0 gets merged with c1.