filter out cases where branches contain voronoi points that "jump" corridors
Until now, the branch-to-corridor algorithm was expecting that two
corridors would always be separated by a point that has more than
two borders. This is actually not the case, so we have to check that
a voronoi point is adjacent to the point before him before adding it
to the current corridor.