replace the algorithm that does inside/outside by two D* runs
We now run two D*, one from the goal and one from the start, to
compute the total cost for a single point, and therefore decide
whether it is within the margin or not.

This replaces half-broken approach we had before. Until we replace
D* by a more suitable algorithm.