support curve annotations
The curve annotation filters allow to tag parts of the curves with
symbols. For instance, the StrongEdgeFilter marks the parts of the
corridor boundaries at which there is a physical edge (i.e. a big
difference in height).

Add a new pass in CorridorPlanner, and add support for display
in the vizkit plugin