filter out small high-cost spots
In rough terrain, it happens often that a host of corridors will
be generated because there are very small spots of high cost.

Instead of taking them into account at the level of the long-range
planner, assume that the local motion planner will handle them and
simplify the resulting corridor plan tremendously.