fix some issues with plan simplification
This commit fixes two main issues:

removeUselessCorridorConnection (renamed to
removeRedundantCorridorConnections) was not acting
recursively, i.e. when connections were copied to the corridor,
redundant could have been added and never removed.

When removing connections, the bidirectional status of the corridors
was never updated. This led to having possible merges not being done,
and having useless connections lying around (inbound connections
to a corridor that would be a dead end). Filter all of those
in removeDeadEnds