filter out point turn connections
When corridors are very small and the margin pretty high, it
can happen that the wrong side of a corridor is connected to
the same side of another. The net result is that the conneciton
requires a "point turn" (i.e. you go along the connection and then
have to enter the corridor in 180 turn).

Filter these connections out as they are outliers