ruby: remove special case for endpoints in #join
The special case was fixing the symptoms but not the problem

The problem is that we have so-called "endpoint corridors" that
are singletons. In some cases, these endpoint corridors are
aways from the other corridors, in which case everything is fine.

In other cases, the endpoint corridor "touches" the first corridor
in the path. Since the endpoint corridor has a zero width, it leads
to an error when we join the endpoint's width curve and the corridor
width curve.

However, the issue is that *in this particular case*, the endpoint
corridor is useless, as the "normal" corridor already leads to the
goal. So, just get rid of it !

It fixes a bug in which the median and width curves returned
by path_to_corridor might not have the same starting or end parameter