fix redundant corridor removal in the non-DAG case
removeRedundatCorridorConnections is used in two contexts: before
the convertion to DAG and after it.

In the first case, it must maintain the full connectivity, i.e.
copy the links in a bidirectional way, while in the second case
it does not need to. Only the second case was dealt with so far,
so add a flag in the method call to tell the method to create
bidirectional connections instead of directional ones