filter out possible connections to obstacles
When creating crossroads, it is possible that a dead-end corridor gets
connected to a proper one *at its end*, i.e. that a corridor ends up
having a connection to an obstacle.

Filter them out in markDirections_cost to avoid having some corridors
marked as dead ends while they are not.